feat: Context API를 활용한 전역 상태 관리 및 계층 재구성
