+++ src/components/CardShape.jsx
@@ -1,4 +1,6 @@
 import React, { useState } from 'react';
+import useCardState from '../hooks/useCardState';
+
 import CardCompany from './common/CardCompany';
 import Modal from './common/Modal';
 
@@ -7,28 +9,9 @@ import * as CardShapeStyle from './common/styles/CardShape.style';
 
 const S = { ...CommonStyle, ...CardShapeStyle };
 
-const cardCompanyList = [
-  { color: '#E24141', name: '포코 카드' },
-  { color: '#547CE4', name: '준 카드' },
-  { color: '#73BC6D', name: '공원 카드' },
-  { color: '#DE59B9', name: '브랜 카드' },
-  { color: '#04C09E4F', name: '로이드 카드' },
-  { color: '#E76E9A', name: '도비 카드' },
-  { color: '#F37D3B', name: '콜린 카드' },
-  { color: '#FBCD58', name: '썬 카드' },
-  { color: '#66adff', name: '태태 카드' },
-  { color: '#cc1b32', name: '콜라 카드' },
-  { color: '#69ff66', name: '비녀 카드' },
-  { color: '#000', name: '동키콩 카드' },
-  { color: '#ff5100', name: '꼬재 카드' },
-  { color: '#192dff', name: '블링 카드' },
-  { color: '#f5d742', name: '샐리 카드' },
-  { color: '#1c24ff', name: '호프 카드' },
-  { color: '#b914ff', name: '소피아 카드' },
-  { color: '#1ca0ff', name: '하리 카드' },
-];
+function CardShape() {
+  const { cardNumber, cardOwnerName, dueDate, dimensions, cardCompanyList } = useCardState();
 
-function CardShape({ cardNumber, cardOwnerName, cardDate, dimensions }) {
   const [isShown, setIsShown] = useState(false);
   const [hexColor, setHexColor] = useState('');
   const [cardCompanyName, setCardCompanyName] = useState('');
@@ -55,7 +38,7 @@ function CardShape({ cardNumber, cardOwnerName, cardDate, dimensions }) {
           </S.CardNumber>
           <S.CardInfo>
             <S.CardParagraph width={'120'}>{cardOwnerName || 'NAME'}</S.CardParagraph>
-            <S.CardParagraph>{cardDate || 'MM / YY'}</S.CardParagraph>
+            <S.CardParagraph>{dueDate || 'MM / YY'}</S.CardParagraph>
           </S.CardInfo>
         </S.CardBottom>
       </S.CardBox>
