+++ src/index.jsx
@@ -2,10 +2,13 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import './index.css';
 import App from './App';
+import { CardProvider } from './contexts';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
-  <React.StrictMode>
+  // <React.StrictMode>
+  <CardProvider>
     <App />
-  </React.StrictMode>,
+  </CardProvider>,
+  // </React.StrictMode>,
 );
